WIP: Add support for formatting of EffectiveModelContext to organization tree
Added utility which prints organization tree for a given EffectiveModelContext. Implementation uses pyang (https://github.com/mbj4668/pyang/blob/master/pyang/plugins/tree.py) as guideline, but adheres to RFC8340 where pyang and RFC specification differ.
As there is no specification for nested column alignment (for example, for types of CaseStatement sub-statements), the column alignment was implemented to match the format of the organization tree in RFC8341 for the non-nested statements.

JIRA: YANGTOOLS-1464
Signed-off-by: matus.matok <matus.matok@pantheon.tech>
Change-Id: Ib0c1fa95102cd5409a0daefe088a6b45d92e9817