Add test cases for YANGTOOLS-1474
Parsing of YANG models may fail with the following necessary conditions:
- YANG models include augmentation with an empty grouping
- Parsing is done with no supported feature

The problem manifests itself when the models are processed in order
foo.yang, bar.yang -- but not when bar.yang is processed before
foo.yang.

Unfortunately the order of processing is not predictable in
BuildGlobalContext, hence we repeat the test 4 times.

JIRA: YANGTOOLS-1474
Change-Id: Ib3d55fb92d5a6d34230722ab1cd39e8f8831b6fe
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>