Add test cases for YANGTOOLS-1474
Parsing of YANG models may fail with the following necessary conditions:
- YANG models include augmentation with an empty grouping
- Parsing is done with no supported feature

The test cases attempt to create effective model context for the same
YANG files - the first two are the same tests, and the last two are the
same and the only difference between the two pairs are the order of YANG
files. However, only some of the test cases fail and it's unpredictable
which one will fail.

JIRA: YANGTOOLS-1474
Change-Id: Ib3d55fb92d5a6d34230722ab1cd39e8f8831b6fe
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>