Add test case for YANGTOOLS-1480
Parsing of data models where multiple YANG modules target a YANG path
defined using a grouping with deviations/augmentations.

Add a test case to reproduce the issue. With the test models it appears
that there are a couple of code paths which end up with a parsing error,
so the test is repeated to show the errors.

JIRA: YANGTOOLS-1480
Change-Id: I4d9c6e6361fe2d5383e6de43e3c1bb54bac4e935
Signed-off-by: Sangwook Ha <sangwook.ha@verizon.com>