Clean up YangToSourcesMojo
We have a weird combination of injected and instantiated processors
here, which really serves only for a very bad UT. Since we are testing
YangToSourcesMojo through integration tests, we can forgo all of this.

This leads to a more straightforward execution, as now the Mojo acts
only as a thin interface towards maven. All the logic is contained in
Processor.

Change-Id: Ia64e69ed6a44e3052f43dcc978e066b080928e73
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>