YangTextSchemaContextResolver needs to we aware of supported features
Added new method YangTextSchemaContextResolver.registeredFeatures() which registers supported features for YangTextSchemaContextResolver.
The logic works as follows. If no features were registered, all features are supported. If a set of features for a module is registered, all the features contained are supported. If multiple sets are registered, a union of contained features is supported. A module which has no features registered has all features supported. If there is only an empty set for a module registered, no features are supported in that module.
This method returns a Registration that upon closing removes a support for the given set of features on which the Registration was returned. If a single feature is registered by multiple method calls, removal of the support for such feature requires all the respective Registrations to close.

JIRA: YANGTOOLS-1463
Change-Id: I443d42a6859ce97f238117fabdb7bbdd42488ee5
Signed-off-by: matus.matok <matus.matok@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>