YangTextSchemaContextResolver needs to we aware of supported features
Added new method YangTextSchemaContextResolver.registeredFeatures() which registers supported features for YangTextSchemaContextResolver.
The logic works as follows. If no features were registered, all features are supported. If a set of features is registered, all the features contained are supported. If multiple sets are registered, a union of contained features is supported.
This method returns a Registration that upon closing removes a support for the given set of features on which the Registration was returned. If a single feature is registered by multiple method calls, removal of the support for such feature requires all the respective Registrations to close.
If there is only an empty set registered, no features are supported.

JIRA: YANGTOOLS-1464
Signed-off-by: matus.matok <matus.matok@pantheon.tech>
Change-Id: I443d42a6859ce97f238117fabdb7bbdd42488ee5