Promote FeatureSet
We are using only Set.contains() to evaluate feature expressions and we
need more flexibility that that. FeatureSet already provides this
capability, but unfortunately it also implies all the Set semantics.

This patch refactors FeatureSet to be a simple class, without a tie to
Set and updates IfFeatureExpr to integrate with it.

JIRA: YANGTOOLS-1504
Change-Id: Ie719f7a3d006f2d6ed51c70604236e4105d8f8ed
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>