Disconnect StatementSupportBundle/NamespaceBehaviour.Registry
StatementSupportBundle is a very concrete class and
NamespaceBehaviour.Registry is a general interface. There is a
connection between them in that other Registry implementations are using
StatementSupportBundle -- but other than that, there is just the
self-similar binding to StatementSupportBundle.parent.

Disconnect these two classes to allow them to go their separate ways.

JIRA: YANGTOOLS-1498
Change-Id: I177b11def897f15b1be7e251d2f1712386a102f8
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>