Refactor global NamespaceStorage access
There is only one way we are passing BuildGlobalContext, hence there is
no reason to indirect access through GlobalStorageAccess, as there is
only one implementation of it -- which always has the result handy.

Remove GlobalStorageAccess and introduce GlobalStorage as a
specialization of NamespaceStorage -- and pass that to
NamespaceBehaviour. This also improves coupling a bit, as there is only
a single implementation (as opposed to NamespaceStorage, which has
multiple implementations).

JIRA: YANGTOOLS-1497
Change-Id: Ica03ac855b62541ec5c52ce24044387199c75e9c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>