Enrich YangParserTestUtils with parseYang(String)
Added a new static method parseYang() to YangParserTestUtils which allows for
parsing a yang file directly from s string input. This allows for higher
flexibility as we do not need to deal with resource accessibility across
modules. Additionally, it is sometimes more convenient to prefer this new form
of input rather than the old one.

JIRA: YANGTOOLS-1483
Signed-off-by: matus.matok <matus.matok@pantheon.tech>
Change-Id: I4865b73ba8770669a54dd5be167b2ca2d63f3585