Add NormalizedYangData
RFC8040 yang-data extension is unlike any other piece of data we
normalize, as it has its own identifier -- which does not match
UnresolvedQName.Unqualified, but rather is captured in YangDataName.

This forms the first such piece of data, but realistically we have a
number of others -- like datastore root, RPC input, RPC output,
notification -- all of which are now overloaded to ContainerNode.

For now introduce NormalizedData as the construct above NormalizedNode,
which does not imply a PathArgument being present -- i.e. pulling
contract(), name() and body() from NormalizedNode.

Add NormalizedYangData, which has its own contract and has a single
DataContainerChild as its body.

JIRA: YANGTOOLS-1472
Change-Id: I7ba989b181ac1bdab8117053233c58f5d40a084f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>