WIP: Disconnect NormalizedNode from Identifiable
Using getIdentifier() is quite ugly in that it very much obfuscates
what is going on. Furthermore we may end up moving this method and
having an explicit tie-in with Identifier is hurtful -- we may want to
create a separate interface for that.

Remove reliance on Identifiable and rename the method to pathArgument().

// FIXME: keep getIdentifier() as deprecated default method, along with
//        its overrides,

Change-Id: I5a1f8ca5150c105775dccbf1c21ee9da6908bc10
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>