Replace the use of CopyableNode in generators
The information, about whether a specific statement
was introduced by augmentation, can be derived by traversing
the schema tree in an opposite way - via parents.

This change does exactly that, walking the tree of generators
by parents towards the root (ModuleGenerator) and checks along
the way, whether the associated statement is target of an augmentation.

JIRA: YANGTOOLS-1657
Change-Id: I551c6aa81db0961e5505d58e8a0ce1aec30aeb8d
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>