Test/Reproduce behavior described in MDSAL-713
A (possibly) faulty behavior was found during the work on MDSAL-713
described in the JIRA issue. The problem is in inconsistent reports
of statements being introduced by augmentation. While substatements
of DeclaredCaseEffectiveStateent, that was introduced by augmentation,
are reported as being introduced by augmentation,
the substatements of UndeclaredCaseEffectiveStatemet, which was
introduced by augmentation, are NOT reported that way.

This change introduces a test that aims to reproduce such behavior.

JIRA: MDSAL-713
Change-Id: I97f7e5aad613c2ee11f60935af9d5db5f352fdea
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>