Rework isAugmenting DataContainerStreamerGenerator
This patch introduces a new logic to determine whether substatements,
of the given `schema`, were introduced by augmentation.

The only time this check is needed to be performed is if the `schema`
is a target of an augmentation (1 or more). This is because
of the nature of the EffectiveStatement retrieval from a binding class
through a RuntimeType.

That statement was either regularly defined in the yang file
and we need to check its effective substatements
(whether they came from augmentation) or that statement is one
that was defined within an augment statement in the yang file
and is in SchemaContext within a AugmentEffectiveStatementImpl
(because of the relation between binding class, runtime type
and the effective statement), which means this is was not introduced
by augmentation neither its children were.

JIRA: YANGTOOLS-1657
Change-Id: Iafd269de72ddc4ffb3cd022a7688c875a87c7bc0
Signed-off-by: Šimon Ukuš <simon.ukus@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>