WIP: Add minidom XML representation
YANG has an 'anydata' escape hatch, which allows for transporting
unknown, but known-to-be-YANG-modeled, part of XML Infoset to be
transported inline.

Historically we have relied on org.w3c.dom to model this data, but that
object model is mutable and more generic than we need.

This patch introduces yang-xml, which provides a more suitable object
model, along with two implementations:
- one wrapping a org.w3c.dom.Document
- one built through Document.Builder et al.

// FIXME: add an issue
// FIXME: rehost to its own artifact

Change-Id: I2e50e03724b5bf5013d29ed79013aa4958fb7e75
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>