WIP: Seal NormalizedNode hierarchy
This patch ensures that all of NormalizedNode interface hierarchy is
sealed, so users can rely on instanceof exhaustiveness.

As a side-effect of this change, we promote
node.spi.AbstractNormalizedNode into the API package and create a mirror
class hierarchy around it.

This hierarchy is then extende by the immutable implementations. This
adds a bit of code duplication, but it also makes it very clear what the
expectations are.

We have flushed out a couple of bugs in the process -- like
LeafSetEntryNode not protecting byte[] children.

JIRA: YANGTOOLS-1259
Change-Id: I0764278cff74a482c74381e6b5e4787ffc3f87e1
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>