Remove StmtContext.Mutable.addRequiredSource()
This method is part of our MagicallyDiscoverWhatSourcesWeNeed(tm)
machinery, dating back to 2017.

In most contexts we have a YangLibModuleSet (or equivalent) information
available -- and hence the reactor requires this set to be resolved and
uses it as its primary API.

In contexts that we do not, we are essentially providing adaptation from
three things we observe:
- modules we need to understand
- modules those modules need
- some sources of YANG text (or similar), with various degrees of
  certainty as to what they actually contain

This essentially boils down to re-doing what YangModelDependencyInfo
already does -- except we need to efficiently handle YIN sources as
well.

// FIXME: review the above
// FIXME: more on the implementation side

JIRA: YANGTOOLS-XXX
Change-Id: Ie241f8489803f271958a3238b43d9d53ec9d8bcc
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>