WIP: Refactor NormalizedNodeWriter
We have a downstream user in RESTCONF implementation, which essentially
copy&pastes our logic and adds various sugar on top of that. This patch
refactors NormalizedNodeWriter to make the life of that subclass easier.

The first is centralization of NormalizedNode type dispatch to protected
methods, so that subclasses can piggy-back on those methods without
having to override the dispatch.

The second is introduction of MapBodyOrder as a separate concern
internal to the implementation. This means we do not have two distinct
implementations differing in a single method -- and thus subclasses can
provide a single specialization and select the appropriate ordering for
it.

While we are here we also change various other aspects of this @Beta
class -- notably return values of methods, as we no longer need the
magic boolean which was inherent in the split-dispatch approach.

// FIXME: mountpoint aware?
// FIXME: add an issue

Change-Id: Ife85567b3c5f0ae262f3938e8d3238f3f42f233e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>