Fix NormalizedNodeStreamWriterStack root handling
When we are attempting to startListEntry(), we really need to be looking
also at the root, as we may have been instantiated at the list itself.

Introduce a dedicated method to return that and use it also when
checking for choices.

JIRA: YANGTOOLS-1575
Change-Id: If4f4afa758d9edd358e9a8e7f1f7665b0156d8b2
Signed-off-by: tobias.pobocik <tobias.pobocik@pantheon.tech>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 33d0cd091722a0d21b99c7c52c8087fa45dc0338)